+++ changelog.md
@@ -1,5 +1,5 @@
-# HoyoToon 0.2.2
-
+# HoyoToon 0.2.3
 
 ## API
-- Debug commands for Users System
+- Remove User ScriptableObject storage and related code.
+- Removed Debug Commands